loggin g and more delay refactoring.

+++ EDGAR-d-MFP-lists/job-download-MFP-files.sh
@@ -19,6 +19,9 @@ while [ -n "$1" ]; do
         forceRun=true
         #echo "forceRun=$forceRun"
         ;;
+    "--missingReport")
+        missingReport=true
+        ;;
     "--pubDelay")
         pubDelayHours="$2"
         #echo "pubDelayHours=$pubDelayHours"
@@ -43,30 +46,32 @@ getCount=0
 listCount=0
 for list in $MFPListsDir/*.json; do
     #echo "Processing $list"
-    cat "$list" | jq -r '.[] | [.cik,.accessionNumber,.url,.filingDate] | @csv' | sed -e 's/"//g' |
-        while read -r entry; do
-            cik=$(echo $entry | cut -d, -f1)
-            accessionNumber=$(echo $entry | cut -d, -f2)
-            url=$(echo $entry | cut -d, -f3)
-            filingDate=$(echo $entry | cut -d, -f4)
-            #echo "filingDate=$filingDate"
-            targetDir="$MFPFilesDir/$cik"
-            [ -d "$targetDir" ] || mkdir -p "$targetDir"
-            targetFile="$targetDir/$filingDate-$accessionNumber.xml"
-            if [ ! -s "$targetFile" ]; then
-                echo "download $cik : $accessionNumber"
+    # this next line is moved to end to not create a sub-shell via piping.
+    #cat "$list" | jq -r '.[] | [.cik,.accessionNumber,.url,.filingDate] | @csv' | sed -e 's/"//g' |
+    while read -r entry; do
+        cik=$(echo $entry | cut -d, -f1)
+        accessionNumber=$(echo $entry | cut -d, -f2)
+        url=$(echo $entry | cut -d, -f3)
+        filingDate=$(echo $entry | cut -d, -f4)
+        #echo "filingDate=$filingDate"
+        targetDir="$MFPFilesDir/$cik"
+        [ -d "$targetDir" ] || mkdir -p "$targetDir"
+        targetFile="$targetDir/$filingDate-$accessionNumber.xml"
+        if [ ! -s "$targetFile" ]; then
+            echo "$getCount: download $cik : $accessionNumber"
+            if [ -z "$missingReport" ]; then
                 ../bin/getEDGAR.sh "$url" >"$targetFile"
                 sleep 1
                 if grep '<title>SEC.gov | File Unavailable</title>' $targetFile; then
                     echo "failed download: $targetFile"
                     rm $targetFile
                 fi
-                getCount=$(($getCount + 1))
             fi
-            [ $getCount -gt 19 ] && exit 1
-        done
+            getCount=$(($getCount + 1))
+        fi
+        [ -z "$missingReport" ] && [ $getCount -gt 19 ] && exit 1
+    done < <(cat "$list" | jq -r '.[] | [.cik,.accessionNumber,.url,.filingDate] | @csv' | sed -e 's/"//g')
     listCount=$(($listCount + 1))
-    # limit # of CIKs processed each time.
-    #[ $listCount -gt 20 ] && exit 1
 done
+[ ! -z "$missingReport" ] && echo "$listCount CIKs, $getCount files to be downloaded."
 exit 0
